move reconciler stuff into separate file
